+++ priv/templates/brando.install/migrations/20150520211010_create_pages.exs
@@ -12,6 +12,7 @@ defmodule <%= application_module %>.Repo.Migrations.CreatePages do
       add :status,            :integer
       add :parent_id,         references(:pages), default: nil
       add :creator_id,        references(:users)
+      add :css_classes,       :text
       add :meta_description,  :text
       add :meta_keywords,     :text
       timestamps